+++ awl/tests/models.py
@@ -59,7 +59,7 @@ class VehicleMake(models.Model):
 
 class VehicleModel(models.Model):
     name = models.CharField(max_length=20)
-    year = models.PositiveIntegerField()
+    year = models.PositiveIntegerField(blank=True, null=True)
     vehiclemake = models.ForeignKey(VehicleMake, null=True, blank=True,
         on_delete=models.CASCADE)