+++ extras/sample_site/app/models.py
@@ -9,7 +9,8 @@ class Writer(models.Model):
 
 class Show(models.Model):
     title = models.CharField(max_length=50)
-    writer = models.ForeignKey(Writer, on_delete=models.CASCADE)
+    writer = models.ForeignKey(Writer, on_delete=models.CASCADE, null=True,
+        blank=True)
 
     def __str__(self):
         return 'Show(id=%s %s)' % (self.id, self.title)
@@ -17,8 +18,8 @@ class Show(models.Model):
 
 class Episode(models.Model):
     name = models.CharField(max_length=50)
-    show = models.ForeignKey(Show, on_delete=models.CASCADE)
+    show = models.ForeignKey(Show, on_delete=models.CASCADE, null=True,
+        blank=True)
 
     def __str__(self):
-        return 'Episode(id=%s %s:%s)' % (self.id, self.show.title, 
-            self.name)
+        return 'Episode(id=%s %s:%s)' % (self.id, self.show.title, self.name)